Silence a warning about an undefined variable
  a = undefined_variable_name
is way too obvious for static analysis tools to detect; if we want
to raise an exception, use something rarer.

This doesn’t change behavior of the test code; the log in /tmp/log
handles both NameError and ZeroDivisionError in the same way.

Eventually, we might end up just with an explicit
  raise SomeExceptionType
but that’s a less realistic test code, so let’s not do it just now.