Drop po/Makefile.in.in from git
It is generated same as */Makefile.in and should be treated equally.

In practical terms, this being a symlink doesn’t work well across
merges/rebases for me (tends to be replaced with a regular file
that contains a path the symlink pointed to), and dropping it from
git is I suppose a workaround.