Change async naming conventions.
Establish a _future suffix convention for functions returning futures.

Drop async_ from names of all functions within the async module as
unnecessarily redundant.  To avoid confusion, document that this
module should always be used as “async.function_name” (i.e. each time
referring to the async nature of the functionality).  Hence, also
replace __all__ because it is a maintenance burden, and only useful for
“from async import *”, which we actively want to break rather than
support.