Enable logging of subprocess output
In the original approach, it was impossible to monitor the
progress of a subprocess because all of the stdout/stderr was
being buffered for retrieval only after the process exited.

With this patch, we will now drop stdout into log.debug1() and
stderr into log.error() as they come in (in addition to storing
the complete results for retrieval later). This will make
debugging efforts much simpler.