Add support for systemd targets
This patch causes new systemd unit files to be generated when a
role is deployed.

Target: This unit file will ensure that all dependent services are
started when the role is launched.

Extension units: These units extend the existing dependent
services and provide the reverse of the above: they ensure that
stopping the role target stops all of the dependent processes. It
also guarantees that a failure of one of these processes updates
the rolekit state for the role.

Failure unit: This is a special unit that emits a DBUS message to
roled if one of the dependent services has failed.