Need a D-BUS routine to be notified of unit failure by systemd
The following command will notify roled that there is could be an error:

 dbus-send --system --dest=org.fedoraproject.rolekit1 \
    /org/fedoraproject/rolekit1 \
    org.fedoraproject.rolekit1.NotifyUnitFailed? \
    string:'rolename' \
    string:'instancename'

roled verifies if there is an error by getting the state from the unit target and moves the instance into error state in this case.

Also fixes the dbus-send call to use the --system option. Otherwise the session bus will be used.

New target_unit_state in dbus_utils to get the state form a target unit.