rolectl: Automatically stop services when decommissioning
Prior to this, if the service was still running, we would get an
error telling us that the service couldn't be decommissioned
unless it was in the "ready-to-start" or "error" state.

Fixes https://github.com/libre-server/rolekit/issues/23

Testing Done:
Decommissioned the memcache role while it was still running.

Reviewed at https://reviewboard-fedoraserver.rhcloud.com/r/171/