Add modified configparser for systemd unit files
This is a temporary measure; this code will be going upstream into python-systemd in the near future, but for now we will carry it in rolekit.

There is one known limitation to this implementation; It does not support reading unit files with multiline values. This won't be an issue in rolekit, but it's preventing it from going upstream to python-systemd at the moment.

Testing Done:
Testing performed manually by loading the python module and writing out some sample unit files. Also tested with the kickstart patch.

Reviewed at https://reviewboard-fedoraserver.rhcloud.com/r/178/