New sanitize (settings) for role instances
Sanitize can be used to for example forget passwords of role instances or
to reset other sensitive data from the settings.

The sanitization is done in the role, as rolekit does not know which
settings need to be reset. The role needs to implement the do_sanitize
method for this.

This is currently implemented for password in the databaseserver role and
for dm_password in the domaincontroller role.

Fixes https://github.com/libre-server/rolekit/issues/33

Bugs closed: 33

Reviewed at https://reviewboard-fedoraserver.rhcloud.com/r/197/