Domain Controller: Use instance name for the domain
This patch changes the default behavior of the domain controller
role. After discussions with UXD experts, it was determined that
the domain should never be auto-detected from the environment,
since it is very difficult to change later. It should always come
from some user action.

This patch changes the default domain to use the role instance name
for the domain name, as long as the instance name contains valid
characters. It also changes the behavior of the host_name option to
set the hostname to dc-<16 characters> if it is currently set to
localhost.

Fixes https://github.com/libre-server/rolekit/issues/43

Testing Done:
Deployed multiple domain controllers with varying initial hostnames.

Bugs closed: 43

Reviewed at https://reviewboard-fedoraserver.rhcloud.com/r/203/