clean up extension unit files on decommission
https://github.com/libre-server/rolekit/issues/16

Testing Done:
Deployed databaseserver, checked that extension unit file paths are stored in role JSON file, decommissioned it again and checked that the files are installed, and removed again.

NB: This is only about extension units, the patch for the others is reviewed in #216.

Reviewed at https://reviewboard-fedoraserver.rhcloud.com/r/212/