Update Vagrantfile for Fedora 23
Also add workaround for docker bug

The docker bug is that in many cases (including the Vagrant case), Docker 1.8.2 (the current version in the stable repository) will fail to initialize on first startup. So this patch forces us to use the known-good 1.7.0 from the Fedora 23 frozen package set in order to do the first-time initialization, then we update to the latest version. This part of the patch can be reverted once 1.9.1 lands in the stable repository, since it seems to fix the issue.

Testing Done:
Built and tested changes for reviews 226-231 with `vagrant rsync && vagrant provision` using this new vagrant box. It worked properly.

Reviewed at https://reviewboard-fedoraserver.rhcloud.com/r/232/