Disable units after stopping
If we disable the units before we call stop(), then the RequiredBy
linkage is already removed. This means that any unit that should
be controlled by the role-*.target won't actually be sent the
stop() command.

The correct thing to do here is to stop the role unit first, then
disable it.

Reviewed at https://reviewboard-fedoraserver.rhcloud.com/r/241/